+++ scripts/run_benchmark/run_full_seqeracloud.sh
@@ -14,8 +14,7 @@ publish_dir="s3://openproblems-data/resources/task_cyto_batch_integration/result
 
 # write the parameters to file
 cat > /tmp/params.yaml << HERE
-# input_states: s3://openproblems-data/resources/task_cyto_batch_integration/datasets/**/state.yaml
-input_states: s3://openproblems-data/resources_test/task_cyto_batch_integration/**/state.yaml
+input_states: s3://openproblems-data/resources/task_cyto_batch_integration/datasets/**/state.yaml
 rename_keys: 'input_unintegrated_censored:unintegrated_censored;input_unintegrated:unintegrated;input_validation:validation'
 output_state: "state.yaml"
 publish_dir: "$publish_dir"
@@ -26,7 +25,6 @@ tw launch https://github.com/openproblems-bio/task_cyto_batch_integration.git \
   --pull-latest \
   --main-script target/nextflow/workflows/run_benchmark/main.nf \
   --workspace 53907369739130 \
-  --compute-env 7gRyww9YNGb0c6BUBtLhDP \
   --params-file /tmp/params.yaml \
   --entry-name auto \
   --config common/nextflow_helpers/labels_tw.config \